refactor(cli): rename consumer short-term cache
